Style: 줄간격 통일, 필요없는print함수 삭제, 파일분리
